fixed changing visitor status/ image file query

+++ app/Http/Controllers/API/VisitorsController.php
@@ -60,7 +60,6 @@ class VisitorsController extends BaseController
     public function store(VisitorsRequest $request)
     {
         $buildingRefID = Building::where('qr_id', $request->building_ID)->first()->id;
-        $approval = VisitTypes::where('')
 
         $profile_link = "";
         if($request->profilePhoto) {
@@ -130,7 +129,7 @@ class VisitorsController extends BaseController
             \Image::make($profile_binary)->fit(200, 200)->save('uploads/profiles-visitor/'.$profile_link)->destroy();
         }
 
-        else if(('uploads/profiles-visitor/'.$data->profilePhoto) != null){
+        else if(('uploads/profiles-visitor/'.$data->profilePhoto) != null && $data->profilePhoto != $request->params['data']['profilePhoto']){
             unlink('uploads/profiles-visitor/'.$data->profilePhoto);
             $profile_link = time().'.' . explode('/', explode(':', substr($profile_binary, 0, strpos($profile_binary, ';')))[1])[1];
             \Image::make($profile_binary)->fit(200, 200)->save('uploads/profiles-visitor/'.$profile_link)->destroy();
@@ -141,7 +140,7 @@ class VisitorsController extends BaseController
             \Image::make($frontID_binary)->fit(200, 200)->save('uploads/frontID/'.$frontID_link)->destroy();
         }
 
-        else if(('uploads/frontID/'.$data->front_id) != null){
+        else if(('uploads/frontID/'.$data->front_id) != null && $data->front_id != $request->params['data']['front_id']){
             unlink('uploads/frontID/'.$data->front_id);
             $frontID_link = time().'.' . explode('/', explode(':', substr($frontID_binary, 0, strpos($frontID_binary, ';')))[1])[1];
             \Image::make($frontID_binary)->fit(200, 200)->save('uploads/frontID/'.$frontID_link)->destroy();
@@ -152,7 +151,7 @@ class VisitorsController extends BaseController
             \Image::make($backID_binary)->fit(200, 200)->save('uploads/backID/'.$backID_link)->destroy();
         }
 
-        else if(('uploads/backID/'.$data->back_id) != null){
+        else if(('uploads/backID/'.$data->back_id) != null && $data->back_id != $request->params['data']['back_id']){
             unlink('uploads/backID/'.$data->back_id);
             $backID_link = time().'.' . explode('/', explode(':', substr($backID_binary, 0, strpos($backID_binary, ';')))[1])[1];
             \Image::make($backID_binary)->fit(200, 200)->save('uploads/backID/'.$backID_link)->destroy();
@@ -167,6 +166,7 @@ class VisitorsController extends BaseController
             'profilePhoto' => $profile_link,
             'front_id' => $frontID_link,
             'back_id' => $backID_link,
+            'status' => $request->params['data']['status'],
             'policy' => $request->params['data']['policy'],
         ]);
